requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-google-genai
google-generativeai
chromadb
faiss-cpu
rank-bm25
sentence-transformers
beautifulsoup4
requests
tqdm
rich
streamlit
pandas
python-dotenv
pydantic
loguru
pymupdf
fastapi
aiogram

